+++ src/com/dipasquale/ai/rl/neat/genotype/GenomeCompatibilityCalculator.java
@@ -2,5 +2,5 @@ package com.dipasquale.ai.rl.neat.genotype;
 
 @FunctionalInterface
 public interface GenomeCompatibilityCalculator {
-    float calculateCompatibility(GenomeDefault genome1, GenomeDefault genome2);
+    double calculateCompatibility(GenomeDefault genome1, GenomeDefault genome2);
 }
